+++ app/controllers/user/projects_controller.rb
@@ -1,27 +1,21 @@
-class ProjectsController < ApplicationController
+class User::ProjectsController < ApplicationController
   after_action :verify_authorized
-  after_action :verify_policy_scoped, only: :index
   rescue_from ActiveRecord::RecordNotFound, with: :not_found_error
 
-  def index
-    authorize Project
-    @pagy, @projects = pagy(policy_scope(Project.in_progress.includes(:assigned_to)))
+  def in_progress
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.assigned_to(current_user).in_progress.includes(:assigned_to))
 
     pre_fetch_establishments(@projects)
     pre_fetch_incoming_trusts(@projects)
   end
 
-  def unassigned
-    authorize Project
-    @pagy, @projects = pagy(policy_scope(Project.unassigned_to_user.assigned_to_regional_caseworker_team))
+  def completed
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.assigned_to(current_user).completed)
 
     pre_fetch_establishments(@projects)
     pre_fetch_incoming_trusts(@projects)
-  end
-
-  def show
-    @project = Project.find(params[:id])
-    authorize @project
   end
 
   private def pre_fetch_establishments(projects)